Rename pkg Errors -> MyErrors
