+++ internal/Authorization/Application/Application.go
@@ -2,7 +2,7 @@ package Application
 
 import (
 	Authorization2 "2021_2_GORYACHIE_MEKSIKANSI/internal/Authorization"
-	errPkg "2021_2_GORYACHIE_MEKSIKANSI/internal/Errors"
+	errPkg "2021_2_GORYACHIE_MEKSIKANSI/internal/MyErrors"
 	"2021_2_GORYACHIE_MEKSIKANSI/internal/Interfaces"
 	"2021_2_GORYACHIE_MEKSIKANSI/internal/Utils"
 )
