+++ internal/Authorization/Orm/ORM.go
@@ -3,7 +3,7 @@ package Orm
 import (
 	"2021_2_GORYACHIE_MEKSIKANSI/internal/Authorization"
 	"2021_2_GORYACHIE_MEKSIKANSI/internal/Authorization/Application"
-	errPkg "2021_2_GORYACHIE_MEKSIKANSI/internal/Errors"
+	errPkg "2021_2_GORYACHIE_MEKSIKANSI/internal/MyErrors"
 	"2021_2_GORYACHIE_MEKSIKANSI/internal/Interfaces"
 	"2021_2_GORYACHIE_MEKSIKANSI/internal/Utils"
 	"context"
