+++ internal/Middleware/Api/API.go
@@ -1,7 +1,7 @@
 package Api
 
 import (
-	errPkg "2021_2_GORYACHIE_MEKSIKANSI/internal/Errors"
+	errPkg "2021_2_GORYACHIE_MEKSIKANSI/internal/MyErrors"
 	interfaces "2021_2_GORYACHIE_MEKSIKANSI/internal/Interfaces"
 	"2021_2_GORYACHIE_MEKSIKANSI/internal/Utils"
 	"github.com/valyala/fasthttp"
