+++ internal/MyErrors/CheckErrorsMIddleware.go
@@ -1,17 +1,17 @@
-package Errors
+package MyErrors
 
 import (
 	"encoding/json"
 	"net/http"
 )
 
-func (c *CheckError) CheckErrorRestaurant(err error) (error, []byte, int) {
+func (c *CheckError) CheckErrorCookie(err error) (error, []byte, int) {
 	if err != nil {
 		switch err.Error() {
-		case RGetRestaurantsRestaurantsNotFound:
+		case MGetIdByCookieCookieNotScan:
 			result, errMarshal := json.Marshal(ResultError{
-				Status:  http.StatusNotFound,
-				Explain: RGetRestaurantsRestaurantsNotFound,
+				Status:  http.StatusInternalServerError,
+				Explain: ErrDB,
 			})
 			if errMarshal != nil {
 				c.Logger.Errorf("%s, %v, requestId: %d", ErrMarshal, errMarshal, c.RequestId)
@@ -20,39 +20,38 @@ func (c *CheckError) CheckErrorRestaurant(err error) (error, []byte, int) {
 					},
 					nil, http.StatusInternalServerError
 			}
-			c.Logger.Warnf("%s, requestId: %d", RGetGeneralInfoRestaurantNotFound, c.RequestId)
+			c.Logger.Errorf("%s, requestId: %d", MGetIdByCookieCookieNotScan, c.RequestId)
+			return &Errors{
+					Alias: ErrCheck,
+				},
+				result, http.StatusInternalServerError
+
+		case MGetIdByCookieCookieExpired, MGetIdByCookieCookieNotFound:
+			result, errMarshal := json.Marshal(ResultError{
+				Status:  http.StatusUnauthorized,
+				Explain: err.Error(),
+			})
+			if errMarshal != nil {
+				c.Logger.Errorf("%s, %v, requestId: %d", ErrMarshal, errMarshal, c.RequestId)
+				return &Errors{
+						Alias: ErrMarshal,
+					},
+					nil, http.StatusInternalServerError
+			}
+			c.Logger.Warnf("%s, requestId: %d", err.Error(), c.RequestId)
 			return &Errors{
 					Alias: ErrCheck,
 				},
 				result, http.StatusOK
-
-		case RGetRestaurantsRestaurantsNotScan, RGetRestaurantsRestaurantsNotSelect:
-			result, errMarshal := json.Marshal(ResultError{
-				Status:  http.StatusInternalServerError,
-				Explain: ErrDB,
-			})
-			if errMarshal != nil {
-				c.Logger.Errorf("%s, %v, requestId: %d", ErrMarshal, errMarshal, c.RequestId)
-				return &Errors{
-						Alias: ErrMarshal,
-					},
-					nil, http.StatusInternalServerError
-			}
-			c.Logger.Errorf("%s, requestId: %d", err.Error(), c.RequestId)
-			return &Errors{
-					Alias: ErrCheck,
-				},
-				result, http.StatusInternalServerError
 		}
 	}
 	return nil, nil, IntNil
 }
 
-func (c *CheckError) CheckErrorRestaurantId(err error) (error, []byte, int) {
+func (c *CheckError) CheckErrorAccess(err error) (error, []byte, int) {
 	if err != nil {
 		switch err.Error() {
-		case RGetGeneralInfoRestaurantNotFound, RGetTagsCategoryRestaurantNotScan, RGetMenuDishesNotSelect,
-			RGetDishesRestaurantDishesNotScan, RGetMenuDishesNotFound, RGetTagsTagsNotFound:
+		case MCheckAccessCookieNotScan:
 			result, errMarshal := json.Marshal(ResultError{
 				Status:  http.StatusInternalServerError,
 				Explain: ErrDB,
@@ -64,24 +63,16 @@ func (c *CheckError) CheckErrorRestaurantId(err error) (error, []byte, int) {
 					},
 					nil, http.StatusInternalServerError
 			}
-			c.Logger.Errorf("%s, requestId: %d", err.Error(), c.RequestId)
+			c.Logger.Errorf("%s, requestId: %d", AGeneralSignUpLoginNotUnique, c.RequestId)
 			return &Errors{
 					Alias: ErrCheck,
 				},
 				result, http.StatusInternalServerError
-		}
-	}
-	return nil, nil, IntNil
-}
 
-func (c *CheckError) CheckErrorRestaurantDishes(err error) (error, []byte, int) {
-	if err != nil {
-		switch err.Error() {
-		case RGetDishesDishesNotFound, RGetStructDishesStructDishesNotSelect, RGetStructDishesStructDishesNotScan,
-			RGetRadiosRadiosNotScan, RGetTagsCategoryNotSelect, RGetRadiosRadiosNotSelect, RGetTagsTagsNotFound:
+		case MCheckAccessCookieNotFound:
 			result, errMarshal := json.Marshal(ResultError{
-				Status:  http.StatusInternalServerError,
-				Explain: ErrDB,
+				Status:  http.StatusUnauthorized,
+				Explain: MCheckAccessCookieNotFound,
 			})
 			if errMarshal != nil {
 				c.Logger.Errorf("%s, %v, requestId: %d", ErrMarshal, errMarshal, c.RequestId)
@@ -90,11 +81,11 @@ func (c *CheckError) CheckErrorRestaurantDishes(err error) (error, []byte, int) 
 					},
 					nil, http.StatusInternalServerError
 			}
-			c.Logger.Errorf("%s, requestId: %d", err.Error(), c.RequestId)
+			c.Logger.Warnf("%s, requestId: %d", MCheckAccessCookieNotFound, c.RequestId)
 			return &Errors{
 					Alias: ErrCheck,
 				},
-				result, http.StatusInternalServerError
+				result, http.StatusOK
 		}
 	}
 	return nil, nil, IntNil
