+++ internal/MyErrors/CheckErrorsRestaurant.go
@@ -1,14 +1,32 @@
-package Errors
+package MyErrors
 
 import (
 	"encoding/json"
 	"net/http"
 )
 
-func (c *CheckError) CheckErrorCookie(err error) (error, []byte, int) {
+func (c *CheckError) CheckErrorRestaurant(err error) (error, []byte, int) {
 	if err != nil {
 		switch err.Error() {
-		case MGetIdByCookieCookieNotScan:
+		case RGetRestaurantsRestaurantsNotFound:
+			result, errMarshal := json.Marshal(ResultError{
+				Status:  http.StatusNotFound,
+				Explain: RGetRestaurantsRestaurantsNotFound,
+			})
+			if errMarshal != nil {
+				c.Logger.Errorf("%s, %v, requestId: %d", ErrMarshal, errMarshal, c.RequestId)
+				return &Errors{
+						Alias: ErrMarshal,
+					},
+					nil, http.StatusInternalServerError
+			}
+			c.Logger.Warnf("%s, requestId: %d", RGetGeneralInfoRestaurantNotFound, c.RequestId)
+			return &Errors{
+					Alias: ErrCheck,
+				},
+				result, http.StatusOK
+
+		case RGetRestaurantsRestaurantsNotScan, RGetRestaurantsRestaurantsNotSelect:
 			result, errMarshal := json.Marshal(ResultError{
 				Status:  http.StatusInternalServerError,
 				Explain: ErrDB,
@@ -20,38 +38,21 @@ func (c *CheckError) CheckErrorCookie(err error) (error, []byte, int) {
 					},
 					nil, http.StatusInternalServerError
 			}
-			c.Logger.Errorf("%s, requestId: %d", MGetIdByCookieCookieNotScan, c.RequestId)
+			c.Logger.Errorf("%s, requestId: %d", err.Error(), c.RequestId)
 			return &Errors{
 					Alias: ErrCheck,
 				},
 				result, http.StatusInternalServerError
-
-		case MGetIdByCookieCookieExpired, MGetIdByCookieCookieNotFound:
-			result, errMarshal := json.Marshal(ResultError{
-				Status:  http.StatusUnauthorized,
-				Explain: err.Error(),
-			})
-			if errMarshal != nil {
-				c.Logger.Errorf("%s, %v, requestId: %d", ErrMarshal, errMarshal, c.RequestId)
-				return &Errors{
-						Alias: ErrMarshal,
-					},
-					nil, http.StatusInternalServerError
-			}
-			c.Logger.Warnf("%s, requestId: %d", err.Error(), c.RequestId)
-			return &Errors{
-					Alias: ErrCheck,
-				},
-				result, http.StatusOK
 		}
 	}
 	return nil, nil, IntNil
 }
 
-func (c *CheckError) CheckErrorAccess(err error) (error, []byte, int) {
+func (c *CheckError) CheckErrorRestaurantId(err error) (error, []byte, int) {
 	if err != nil {
 		switch err.Error() {
-		case MCheckAccessCookieNotScan:
+		case RGetGeneralInfoRestaurantNotFound, RGetTagsCategoryRestaurantNotScan, RGetMenuDishesNotSelect,
+			RGetDishesRestaurantDishesNotScan, RGetMenuDishesNotFound, RGetTagsTagsNotFound:
 			result, errMarshal := json.Marshal(ResultError{
 				Status:  http.StatusInternalServerError,
 				Explain: ErrDB,
@@ -63,16 +64,24 @@ func (c *CheckError) CheckErrorAccess(err error) (error, []byte, int) {
 					},
 					nil, http.StatusInternalServerError
 			}
-			c.Logger.Errorf("%s, requestId: %d", AGeneralSignUpLoginNotUnique, c.RequestId)
+			c.Logger.Errorf("%s, requestId: %d", err.Error(), c.RequestId)
 			return &Errors{
 					Alias: ErrCheck,
 				},
 				result, http.StatusInternalServerError
+		}
+	}
+	return nil, nil, IntNil
+}
 
-		case MCheckAccessCookieNotFound:
+func (c *CheckError) CheckErrorRestaurantDishes(err error) (error, []byte, int) {
+	if err != nil {
+		switch err.Error() {
+		case RGetDishesDishesNotFound, RGetStructDishesStructDishesNotSelect, RGetStructDishesStructDishesNotScan,
+			RGetRadiosRadiosNotScan, RGetTagsCategoryNotSelect, RGetRadiosRadiosNotSelect, RGetTagsTagsNotFound:
 			result, errMarshal := json.Marshal(ResultError{
-				Status:  http.StatusUnauthorized,
-				Explain: MCheckAccessCookieNotFound,
+				Status:  http.StatusInternalServerError,
+				Explain: ErrDB,
 			})
 			if errMarshal != nil {
 				c.Logger.Errorf("%s, %v, requestId: %d", ErrMarshal, errMarshal, c.RequestId)
@@ -81,11 +90,11 @@ func (c *CheckError) CheckErrorAccess(err error) (error, []byte, int) {
 					},
 					nil, http.StatusInternalServerError
 			}
-			c.Logger.Warnf("%s, requestId: %d", MCheckAccessCookieNotFound, c.RequestId)
+			c.Logger.Errorf("%s, requestId: %d", err.Error(), c.RequestId)
 			return &Errors{
 					Alias: ErrCheck,
 				},
-				result, http.StatusOK
+				result, http.StatusInternalServerError
 		}
 	}
 	return nil, nil, IntNil
