+++ internal/MyErrors/CheckErrorsRestaurant_test.go
@@ -1,4 +1,4 @@
-package Errors
+package MyErrors
 
 import (
 	test "2021_2_GORYACHIE_MEKSIKANSI/test"
