+++ internal/Order/Api/API.go
@@ -2,7 +2,7 @@ package Api
 
 import (
 	"2021_2_GORYACHIE_MEKSIKANSI/internal/Authorization"
-	errPkg "2021_2_GORYACHIE_MEKSIKANSI/internal/Errors"
+	errPkg "2021_2_GORYACHIE_MEKSIKANSI/internal/MyErrors"
 	interfaces "2021_2_GORYACHIE_MEKSIKANSI/internal/Interfaces"
 	"2021_2_GORYACHIE_MEKSIKANSI/internal/Order"
 	"2021_2_GORYACHIE_MEKSIKANSI/internal/Utils"
