+++ internal/Profile/Orm/ORM.go
@@ -1,7 +1,7 @@
 package Orm
 
 import (
-	errPkg "2021_2_GORYACHIE_MEKSIKANSI/internal/Errors"
+	errPkg "2021_2_GORYACHIE_MEKSIKANSI/internal/MyErrors"
 	"2021_2_GORYACHIE_MEKSIKANSI/internal/Interfaces"
 	"2021_2_GORYACHIE_MEKSIKANSI/internal/Profile"
 	Utils2 "2021_2_GORYACHIE_MEKSIKANSI/internal/Utils"
